+++ jquery.reel.js
@@ -25,7 +25,7 @@
  * jQuery Reel
  * http://jquery.vostrel.cz/reel
  * Version: 1.1.3-devel
- * Updated: 2011-05-08
+ * Updated: 2011-10-21
  *
  * Requires jQuery 1.4.2 or higher
  */
@@ -99,7 +99,8 @@ jQuery.reel || (function($, window, document, undefined){
       vertical:       false, // switches orbital object movie to vertical mode
       wheelable:       true, // mouse wheel interaction (allowed by default)
 
-      attr:              {}  // initial attribute-value pairs map for the IMG tag
+      attr:              {}, // initial attribute-value pairs map for the IMG tag
+      scrollable:      true  // allow page scroll (allowed by default; applies only to touch devices)
     }
     // [deprecated] options defaults may be gone anytime soon
   }
@@ -255,6 +256,7 @@ jQuery.reel || (function($, window, document, undefined){
               $hi= $(_div_tag_, { 'class': hi_klass,
                 css: { position: _absolute_, left: 0, top: 0, width: space.x, height: space.y, background: _hex_black_, opacity: 0 }
               }).appendTo($overlay),
+              scrollable= !get(_reeling_) || opt.rows <= 1 || !opt.orbital || opt.scrollable,
               area= set(_area_, $(opt.area || $hi ))
             if ($.reel.touchy){
               // workaround for downsizing-sprites-bug-in-iPhoneOS inspired by Katrin Ackermann
@@ -265,7 +267,7 @@ jQuery.reel || (function($, window, document, undefined){
               });
               area
                 .bind(_touchstart_, function(e){ t.trigger('down', [finger(e).clientX, finger(e).clientY, true]); })
-                .bind(_touchmove_, function(e){ t.trigger('slide', [finger(e).clientX, finger(e).clientY, true]); return !(opt.rows > 1 || opt.orbital || get(_reeling_)) })
+                .bind(_touchmove_, function(e){ t.trigger('slide', [finger(e).clientX, finger(e).clientY, true]); return !scrollable })
                 .bind(_touchend_, function(e){ t.trigger('up', [true]); return false })
                 .bind(_touchcancel_, function(e){ t.trigger('up', [true]); return false })
             }else{